# schema/src/swagger.yaml
openapi: "3.0.0"
info:
  version: "0.1.0"
  title: whatsup
paths:
  /user:
    get:
      summary: get user profile
      description: ""
      operationId: getUser
      tags:
        - user
      parameters:
        - name: authorization
          in: header
          required: true
          schema:
            type: string
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /user/register:
    post:
      summary: register a user
      description: ""
      operationId: registerUser
      tags:
        - user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegisterData"
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /user/login:
    post:
      summary: login
      description: ""
      operationId: login
      tags:
        - user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserLoginData"
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResult"

  /chats:
    get:
      summary: get all chats
      description: ""
      operationId: listChats
      tags:
        - chat
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chats"
    post:
      summary: create a chat
      description: ""
      operationId: createChat
      tags:
        - chat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatCreateData"
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chat"
  /chats/{id}:
    get:
      summary: get a chat
      description: ""
      operationId: getChat
      tags:
        - chat
      parameters:
        - name: id
          in: path
          description: chat id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chat"
    delete:
      summary: delete a chat
      description: ""
      operationId: deleteChat
      tags:
        - chat
      parameters:
        - name: id
          in: path
          description: chat id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: ""

tags:
  - name: user
    description: "user operations"
  - name: chat
    description: "chat operations"

security:
  - bearer: []

components:
  schemas:
    AuthorizationData:
      type: object
      properties:
        userId:
          type: integer
          format: int64
      required:
        - usedId
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
      required:
        - id
        - username
    UserRegisterData:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    UserLoginData:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    LoginResult:
      type: object
      properties:
        success:
          type: boolean
        token:
          type: string
      required:
        - success
    Chats:
      type: array
      items:
        $ref: "#/components/schemas/Chat"
    Chat:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
      required:
        - id
        - name
    ChatCreateData:
      type: object
      properties:
        name:
          type: string
      required:
        - name

  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT